Refactor mapping and saving logic
